chore: Update package-lock.json and package.json to version 0.39.2
Update the package-lock.json and package.json files to reflect the new version 0.39.2. This update includes changes to dependencies such as "@fnet/cli-project-common" and "@fnet/cli-project-flow". Additionally, the "dts_enabled" feature has been added to the project's features.

Refactor create-dts and init-dependencies

In the create-dts module, a check has been added to skip the creation of .d.ts files if the "dts_enabled" feature is not enabled. In the init-dependencies module, the "rollup-plugin-dts" package is now only added as a devDependency if the "dts_enabled" feature is enabled.

Refactor init-features

The init-features module now sets the "dts_enabled" feature based on the value of the "dts" property in the project's features. If the "dts" property is set to true or if it is an object with an "enabled" property set to true, the "dts_enabled" feature will be enabled.

Refactor lib-builder and wf-builder

Both the lib-builder and wf-builder modules have been updated to include the creation of .d.ts files using the createDts function.